requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.3.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
mysql-connector-python==8.2.0
